Point estimate functions

Helper functions to define point estimates for aggregateReplicates.

WeightedMean(var, na.rm = FALSE)

WeightedMedian(var, na.rm = FALSE)

WeightedSum(var, na.rm = FALSE)

WeightedRatio(condition, na.rm = FALSE)

WeightedSize(condition, na.rm = FALSE)

Arguments

var
A variable name in the dataset to be used for estimation

na.rm
logical. Should missing values (including NaN) be removed?

condition
A string specifying a condition that can be evaluated for each observation. This condition may contain column names in the dataset.

Value

A function taking a subset of the data and a weight vector as input and returning a single numeric value. For example function(subset, w){ sum(w) }

Details

WeightedRatio returns the relative number of observations in each group satisfying condition. For example, a poverty-risk rate for each state.

Examples

df <- data.table::data.table(age = rnorm(100, mean = 30, sd = 10), w = 1)

estimator <- WeightedMean("age")
estimator(df, df$w)

#> [1] 30.70789

estimator <- WeightedRatio("age>30")
estimator(df, df$w)

#> [1] 0.54

estimator <- WeightedSize("age>30")
estimator(df, df$w)

#> [1] 54