Calculate point estimates for weights and bootstrap weights

Calculate an estimate for each group and each bootstrap replicate of survey weights. The replicate weihgts are assumed to be in wide format as returned by recalib and draw.bootstrap

aggregateReplicates(dat, FUN, groupVars = NULL, weights = attr(dat,
  "weights"), b.weights = attr(dat, "b.rep"))

Arguments

dat
A data.table containing bootstrap weights. This will usually be the output of recalib or draw.bootstrap.

FUN
An aggregation function as returned by WeightedMean, WeightedMedian, WeightedSum, WeightedRatio and WeightedSize.

groupVars
Column names in dat refering to grouping variables. Typically, those columns are factors or integers

weights
Column names for the survey weights

b.weights
Column names for the boostrap weights

Examples

eusilc <- surveysd:::demo.eusilc()
dat_boot <- draw.bootstrap(eusilc, REP = 10, hid = "db030", weights = "rb050",
                           strata = c("db040"), period = "year")
dat_boot_calib <- recalib(dat_boot, conP.var = c("rb090", "age"), conH.var = c("db040", "hsize"))

#> Convergence reached in  4  steps 
#> Convergence reached in  5  steps 
#> Convergence reached in  6  steps 
#> Convergence reached in  4  steps 
#> Convergence reached in  5  steps 
#> Convergence reached in  4  steps 
#> Convergence reached in  4  steps 
#> Convergence reached in  3  steps 
#> Convergence reached in  4  steps 
#> Convergence reached in  4  steps 

## poverty rate for each year
aggregateReplicates(dat_boot_calib, WeightedRatio("povmd60 == 1"), c("year"))

#>    year     rb050        w1        w2        w3        w4        w5        w6
#> 1: 2010 0.1444422 0.1460785 0.1448612 0.1344769 0.1524899 0.1507202 0.1444870
#> 2: 2011 0.1461885 0.1484984 0.1488972 0.1410871 0.1537014 0.1465584 0.1469416
#> 3: 2012 0.1495303 0.1497175 0.1549785 0.1473796 0.1554456 0.1499817 0.1501293
#> 4: 2013 0.1522681 0.1548291 0.1558679 0.1474812 0.1565454 0.1529620 0.1556264
#> 5: 2014 0.1556545 0.1582114 0.1576702 0.1511088 0.1531412 0.1614744 0.1590959
#> 6: 2015 0.1568041 0.1584588 0.1609599 0.1514466 0.1509541 0.1617068 0.1611071
#> 7: 2016 0.1553374 0.1584738 0.1589738 0.1521750 0.1515783 0.1586161 0.1586466
#> 8: 2017 0.1527709 0.1531283 0.1550015 0.1514241 0.1457766 0.1533927 0.1576971
#>           w7        w8        w9       w10
#> 1: 0.1434454 0.1399357 0.1491166 0.1375642
#> 2: 0.1431254 0.1448211 0.1513300 0.1420413
#> 3: 0.1483345 0.1540979 0.1545494 0.1466315
#> 4: 0.1498656 0.1543825 0.1571175 0.1519471
#> 5: 0.1554837 0.1575109 0.1588861 0.1563591
#> 6: 0.1599877 0.1583470 0.1607239 0.1617658
#> 7: 0.1546426 0.1549783 0.1607163 0.1596646
#> 8: 0.1492205 0.1557786 0.1562696 0.1533550

## mean income for each year and gender
aggregateReplicates(dat_boot_calib, WeightedMean("py010n", na.rm = TRUE), c("year", "rb090"))

#>     year  rb090   rb050      w1      w2      w3      w4      w5      w6      w7
#>  1: 2010 female 3448055 3446815 3481772 3425930 3352383 3342360 3496223 3528896
#>  2: 2010   male 6894942 6747515 6921692 6992245 7034746 6862637 6869159 6877138
#>  3: 2011 female 3448055 3525009 3449742 3509365 3286975 3361958 3443138 3460091
#>  4: 2011   male 6894942 6741774 6918564 6900303 6943210 6906536 6865388 6885185
#>  5: 2012 female 3448055 3518685 3495445 3456934 3326634 3372711 3501736 3462927
#>  6: 2012   male 6894942 6786386 6872892 6862153 6875790 6944777 6917657 6881153
#>  7: 2013 female 3448055 3470771 3496511 3457106 3348204 3405537 3504521 3490223
#>  8: 2013   male 6894942 6862165 6972606 6866139 6887149 6961499 6918440 6843497
#>  9: 2014 female 3448055 3450082 3535969 3441153 3314689 3394618 3518235 3435958
#> 10: 2014   male 6894942 6859557 6908295 6895244 6911616 6905245 7027490 6900977
#> 11: 2015 female 3448055 3461373 3648256 3367782 3322462 3383058 3518420 3413982
#> 12: 2015   male 6894942 6883104 6843152 6739539 6941650 6896477 6995581 6935441
#> 13: 2016 female 3448055 3429903 3580314 3464551 3359003 3362655 3503749 3414971
#> 14: 2016   male 6894942 6936401 6871674 6760746 6963953 6745548 6982633 6995213
#> 15: 2017 female 3448055 3512271 3578278 3491750 3384864 3360017 3540801 3463467
#> 16: 2017   male 6894942 6897103 6920140 6675080 6925404 6774829 6964029 6801955
#>          w8      w9     w10
#>  1: 3402694 3425743 3376234
#>  2: 6923708 7000542 6885008
#>  3: 3369152 3406140 3372592
#>  4: 6918645 6931614 6921617
#>  5: 3423363 3313153 3344292
#>  6: 6974861 6835562 6903671
#>  7: 3450466 3323874 3405381
#>  8: 7048960 6870778 6909682
#>  9: 3457156 3296647 3393549
#> 10: 6919340 6838303 6922644
#> 11: 3543242 3377294 3318051
#> 12: 6895455 6827437 6995828
#> 13: 3563035 3432349 3381045
#> 14: 6873107 6811026 6921389
#> 15: 3644530 3382359 3437113
#> 16: 6945011 6861692 6882551