Generate new houshold ID for survey data with rotating panel design taking into account split households

Generating a new houshold ID for survey data using a houshold ID and a personal ID. For surveys with rotating panel design containing housholds, houshold members can move from an existing household to a new one, that was not originally in the sample. This leads to the creation of so called split households. Using a peronal ID (that stays fixed over the whole survey), an indicator for different time steps and a houshold ID, a new houshold ID is assigned to the original and the split household.

generate.HHID(dat, period = "RB010", pid = "RB030", hid = "DB030")

Arguments

dat
data table of data frame containing the survey data

period
column name of dat containing an indicator for the rotations, e.g years, quarters, months, ect...

pid
column name of dat containing the personal identifier. This needs to be fixed for an indiviual throught the whole survey

hid
column name of dat containing the household id. This needs to for a household throught the whole survey

Value

the survey data dat as data.table object containing a new and an old household ID. The new household ID which considers the split households is now named hid and the original household ID has a trailing "_orig".

Examples

library(surveysd)
library(laeken)
library(data.table)

eusilc <- surveysd:::demo.eusilc(n=4)

# create spit households
eusilc[,rb030split:=rb030]

#>        db030 hsize         db040   rb030       age  rb090 pl030 pb220a   py010n
#>     1:     1 (2,3]         Tyrol     101   (25,45] female     2     AT  9756.25
#>     2:     1 (2,3]         Tyrol     102   (25,45]   male     1  Other 12471.60
#>     3:     1 (2,3]         Tyrol     103 (-Inf,16]   male  <NA>   <NA>       NA
#>     4:     2 (3,4]         Tyrol     201   (25,45] female     7     AT 12487.03
#>     5:     2 (3,4]         Tyrol     202   (25,45]   male     1     AT 42821.23
#>    ---                                                                         
#> 59304:  7498 (3,4] Lower Austria  749804 (-Inf,16] female     4     AT     0.00
#> 59305:  7499 (0,1] Upper Austria  749901   (25,45] female     1     AT 13962.56
#> 59306:  5999 (0,1]         Tyrol  599901   (25,45]   male     1     AT 14685.18
#> 59307: 10500 (1,2]         Tyrol 1050001   (45,65]   male     1     AT 20606.82
#> 59308: 10500 (1,2]         Tyrol 1050002   (45,65] female     6     AT     0.00
#>        py050n  py090n py100n py110n py120n py130n py140n hy040n  hy050n hy070n
#>     1:      0    0.00      0      0      0      0      0 4273.9 2428.11      0
#>     2:      0    0.00      0      0      0      0      0 4273.9 2428.11      0
#>     3:     NA      NA     NA     NA     NA     NA     NA 4273.9 2428.11      0
#>     4:      0    0.00      0      0      0      0      0    0.0 1549.72      0
#>     5:      0    0.00      0      0      0      0      0    0.0 1549.72      0
#>    ---                                                                        
#> 59304:      0    0.00      0      0      0      0      0    0.0 1955.19      0
#> 59305:      0    0.00      0      0      0      0      0    0.0    0.00      0
#> 59306:      0    0.00      0      0      0      0      0    0.0    0.00      0
#> 59307:      0    0.00      0      0      0      0      0    0.0    0.00      0
#> 59308:      0 3825.63      0      0      0      0      0    0.0    0.00      0
#>        hy080n hy090n hy110n hy130n hy145n eqSS eqIncome    db090    rb050 year
#>     1:      0  33.39      0      0      0  1.8 16090.69 504.5696 504.5696 2010
#>     2:      0  33.39      0      0      0  1.8 16090.69 504.5696 504.5696 2010
#>     3:      0  33.39      0      0      0  1.8 16090.69 504.5696 504.5696 2010
#>     4:      0   2.13      0      0      0  2.1 27076.24 493.3824 493.3824 2010
#>     5:      0   2.13      0      0      0  2.1 27076.24 493.3824 493.3824 2010
#>    ---                                                                        
#> 59304:      0   0.00      0      0      0  2.5 19867.59 556.4260 556.4260 2013
#> 59305:      0 424.85      0      0      0  1.0 14989.37 643.2557 643.2557 2013
#> 59306:      0 120.65      0      0      0  1.0 14805.83 679.7288 679.7288 2013
#> 59307:      0   0.00      0      0      0  1.5 18707.14 567.1544 567.1544 2013
#> 59308:      0   0.00      0      0      0  1.5 27157.23 567.1544 567.1544 2013
#>        povmd60 rb030split
#>     1:       0        101
#>     2:       0        102
#>     3:       0        103
#>     4:       0        201
#>     5:       0        202
#>    ---                   
#> 59304:       0     749804
#> 59305:       0     749901
#> 59306:       0     599901
#> 59307:       0    1050001
#> 59308:       0    1050002

year <- eusilc[,unique(year)]
year <- year[-1]
leaf_out <- c()
for(y in year){
  split.person <- eusilc[year==(y-1)&!duplicated(db030)&!db030%in%leaf_out,
                         sample(rb030,20)]
  overwrite.person <- eusilc[year==(y)&!duplicated(db030)&!db030%in%leaf_out,
                             .(rb030=sample(rb030,20))]
  overwrite.person[,c("rb030split","year_curr"):=.(split.person,y)]

  eusilc[overwrite.person,rb030split:=i.rb030split,on=.(rb030,year>=year_curr)]
  leaf_out <- c(leaf_out,
                eusilc[rb030%in%c(overwrite.person$rb030,overwrite.person$rb030split),
                unique(db030)])
}

# pid which are in split households
eusilc[,.(uniqueN(db030)),by=list(rb030split)][V1>1]

#>     rb030split V1
#>  1:       4301  2
#>  2:      15601  2
#>  3:      80401  2
#>  4:      93801  2
#>  5:     107601  2
#>  6:     137901  2
#>  7:     181701  2
#>  8:     203801  2
#>  9:     235801  2
#> 10:     255101  2
#> 11:     276801  2
#> 12:     285701  2
#> 13:     326501  2
#> 14:     332401  2
#> 15:     342901  2
#> 16:     357601  2
#> 17:     383501  2
#> 18:     419201  2
#> 19:     460601  2
#> 20:     582601  2
#> 21:     602301  2
#> 22:     605201  2
#> 23:     608501  2
#> 24:     623801  2
#> 25:     628501  2
#> 26:     643101  2
#> 27:     645801  2
#> 28:     685501  2
#> 29:     689701  2
#> 30:     694001  2
#> 31:     695101  2
#> 32:     715401  2
#> 33:     721301  2
#> 34:     726301  2
#> 35:     729301  2
#> 36:     732201  2
#> 37:     740401  2
#> 38:     741201  2
#> 39:     745701  2
#> 40:     746501  2
#> 41:     750901  2
#> 42:     756001  2
#> 43:     764601  2
#> 44:     781701  2
#> 45:     782701  2
#> 46:     784201  2
#> 47:     794601  2
#> 48:     802001  2
#> 49:     804501  2
#> 50:     813101  2
#> 51:     817301  2
#> 52:     818301  2
#> 53:     833501  2
#> 54:     834501  2
#> 55:     853401  2
#> 56:     853901  2
#> 57:     866401  2
#> 58:     870101  2
#> 59:     891201  2
#> 60:     894801  2
#>     rb030split V1

eusilc.new <- generate.HHID(eusilc,period="year",pid="rb030split",hid="db030")

# no longer any split households in the data
eusilc.new[,.(uniqueN(db030)),by=list(rb030split)][V1>1]

#> Empty data.table (0 rows) of 2 cols: rb030split,V1