Plot surveysd-Objects

Plot results of calc.stError()

# S3 method for surveysd
plot(x, variable = x$param$var[1],
  type = c("summary", "grouping"), groups = NULL, sd.type = c("dot",
  "ribbon"), ...)

Arguments

x
object of class 'surveysd' output of function calc.stError

variable
Name of the variable for which standard errors have been calcualated in dat

type
can bei either 'summary' or 'grouping', default value is 'summary'. For 'summary' a barplot is created giving an overview of the number of estimates having the flag smallGroup, cvHigh, both or none of them. For 'grouping' results for point estimate and standard error are plotted for pre defined groups.

groups
If type='grouping' variables must be defined by which the data is grouped. Only 2 levels are supported as of right now. If only one group is defined the higher group will be the estimate over the whole period. Results are plotted for the first argument in groups as well as for the combination of groups[1] and groups[2].

sd.type
can bei either 'ribbon' or 'dot' and is only used if type='grouping'. Default is "dot" For sd.type='dot' point estimates are plotted and flagged if the corresponding standard error and/or the standard error using the mean over k-periods exceeded the value cv.limit (see calc.stError). For sd.type='ribbon' the point estimates including ribbons, defined by point estimate +- estimated standard error are plotted. The calculated standard errors using the mean over k periods are plotted using less transparency. Results for the higher level (~groups[1]) are coloured grey.

...
additional arguments supplied to plot.

Examples

# NOT RUN {
library(surveysd)
library(laeken)
library(data.table)

eusilc <- surveysd:::demo.eusilc()

dat_boot <- draw.bootstrap(eusilc,REP=250,hid="db030",weights="rb050",strata=c("db040"),
                           period="year")

# calibrate weight for bootstrap replicates
dat_boot_calib <- recalib(dat=copy(dat_boot),hid="db030",weights="rb050",
                          period="year",b.rep=paste0("w",1:250),
                          conP.var=c("rb090"),conH.var = c("db040"))

# estimate weightedRatio for povmd60 per period
group <- list("rb090","db040",c("rb090","db040"))
err.est <- calc.stError(dat_boot_calib,weights="rb050",b.weights=paste0("w",1:250),
                        period="year",var="povmd60",fun="weightedRatio",
                        group=c("rb090","db040"),period.diff=NULL,period.mean=NULL)

plot(err.est)

# plot results for rb090
# dotted line is the result on the national level
plot(err.est,type="grouping",groups="rb090")

# plot results for rb090 in each db040
plot(err.est,type="grouping",groups=c("rb090","db040"))

# plot results for db040 in each rb090
plot(err.est,type="grouping",groups=c("db040","rb090"))
# }