Generate new houshold ID for survey data with rotating panel design taking into account split households

Generating a new houshold ID for survey data using a houshold ID and a personal ID. For surveys with rotating panel design containing housholds, houshold members can move from an existing household to a new one, that was not originally in the sample. This leads to the creation of so called split households. Using a peronal ID (that stays fixed over the whole survey), an indicator for different time steps and a houshold ID, a new houshold ID is assigned to the original and the split household.

generate.HHID(dat, period = "RB010", pid = "RB030", hid = "DB030")

Arguments

dat
data table of data frame containing the survey data

period
column name of dat containing an indicator for the rotations, e.g years, quarters, months, ect...

pid
column name of dat containing the personal identifier. This needs to be fixed for an indiviual throught the whole survey

hid
column name of dat containing the household id. This needs to for a household throught the whole survey

Value

the survey data dat as data.table object containing a new and an old household ID. The new household ID which considers the split households is now named hid and the original household ID has a trailing "_orig".

Examples

library(surveysd)
library(laeken)
library(data.table)

eusilc <- surveysd:::demo.eusilc()

# create spit households   
eusilc[,rb030split:=rb030]

#>         db030 hsize         db040   rb030       age  rb090 pl030 pb220a
#>      1:     1 (2,3]         Tyrol     101   (25,45] female     2     AT
#>      2:     1 (2,3]         Tyrol     102   (25,45]   male     1  Other
#>      3:     1 (2,3]         Tyrol     103 (-Inf,16]   male  <NA>   <NA>
#>      4:     2 (3,4]         Tyrol     201   (25,45] female     7     AT
#>      5:     2 (3,4]         Tyrol     202   (25,45]   male     1     AT
#>     ---                                                                
#> 118612:  5997 (3,4] Lower Austria  599704 (-Inf,16] female     4     AT
#> 118613: 16499 (0,1] Upper Austria 1649901   (25,45] female     1     AT
#> 118614: 16500 (0,1]         Tyrol 1650001   (25,45]   male     1     AT
#> 118615: 15000 (1,2]         Tyrol 1500001   (45,65]   male     1     AT
#> 118616: 15000 (1,2]         Tyrol 1500002   (45,65] female     6     AT
#>           py010n py050n  py090n py100n py110n py120n py130n py140n hy040n
#>      1:  9756.25      0    0.00      0      0      0      0      0 4273.9
#>      2: 12471.60      0    0.00      0      0      0      0      0 4273.9
#>      3:       NA     NA      NA     NA     NA     NA     NA     NA 4273.9
#>      4: 12487.03      0    0.00      0      0      0      0      0    0.0
#>      5: 42821.23      0    0.00      0      0      0      0      0    0.0
#>     ---                                                                  
#> 118612:     0.00      0    0.00      0      0      0      0      0    0.0
#> 118613: 13962.56      0    0.00      0      0      0      0      0    0.0
#> 118614: 14685.18      0    0.00      0      0      0      0      0    0.0
#> 118615: 20606.82      0    0.00      0      0      0      0      0    0.0
#> 118616:     0.00      0 3825.63      0      0      0      0      0    0.0
#>          hy050n hy070n hy080n hy090n hy110n hy130n hy145n eqSS eqIncome
#>      1: 2428.11      0      0  33.39      0      0      0  1.8 16090.69
#>      2: 2428.11      0      0  33.39      0      0      0  1.8 16090.69
#>      3: 2428.11      0      0  33.39      0      0      0  1.8 16090.69
#>      4: 1549.72      0      0   2.13      0      0      0  2.1 27076.24
#>      5: 1549.72      0      0   2.13      0      0      0  2.1 27076.24
#>     ---                                                                
#> 118612: 1955.19      0      0   0.00      0      0      0  2.5 26508.20
#> 118613:    0.00      0      0 424.85      0      0      0  1.0 26270.99
#> 118614:    0.00      0      0 120.65      0      0      0  1.0 12850.17
#> 118615:    0.00      0      0   0.00      0      0      0  1.5 13548.47
#> 118616:    0.00      0      0   0.00      0      0      0  1.5 24523.01
#>            db090    rb050 year povmd60 rb030split
#>      1: 504.5696 504.5696 2010       0        101
#>      2: 504.5696 504.5696 2010       0        102
#>      3: 504.5696 504.5696 2010       0        103
#>      4: 493.3824 493.3824 2010       0        201
#>      5: 493.3824 493.3824 2010       0        202
#>     ---                                          
#> 118612: 556.4260 556.4260 2017       0     599704
#> 118613: 643.2557 643.2557 2017       0    1649901
#> 118614: 679.7288 679.7288 2017       0    1650001
#> 118615: 567.1544 567.1544 2017       0    1500001
#> 118616: 567.1544 567.1544 2017       0    1500002

year <- eusilc[,unique(year)]
year <- year[-1]
leaf_out <- c()
for(y in year){
  split.person <- eusilc[year==(y-1)&!duplicated(db030)&!db030%in%leaf_out,
                         sample(rb030,20)]
  overwrite.person <- eusilc[year==(y)&!duplicated(db030)&!db030%in%leaf_out,
                             .(rb030=sample(rb030,20))]
  overwrite.person[,c("rb030split","year_curr"):=.(split.person,y)]

  eusilc[overwrite.person,rb030split:=i.rb030split,on=.(rb030,year>=year_curr)]
  leaf_out <- c(leaf_out,
                eusilc[rb030%in%c(overwrite.person$rb030,overwrite.person$rb030split),
                unique(db030)])
}

# pid which are in split households
eusilc[,.(uniqueN(db030)),by=list(rb030split)][V1>1]

#>      rb030split V1
#>   1:       6301  2
#>   2:      29401  2
#>   3:      53501  2
#>   4:      62401  2
#>   5:      70001  2
#>  ---              
#> 136:    1465601  2
#> 137:    1468001  2
#> 138:    1483501  2
#> 139:    1484101  2
#> 140:    1496101  2

eusilc.new <- generate.HHID(eusilc,period="year",pid="rb030split",hid="db030")

# no longer any split households in the data
eusilc.new[,.(uniqueN(db030)),by=list(rb030split)][V1>1]

#> Empty data.table (0 rows) of 2 cols: rb030split,V1