Perform one step of iterative proportional updating

C++ routines to invoke a single iteration of the Iterative proportional updating (IPU) scheme. Targets and classes are assumed to be one dimensional in the ipf_step functions. combine_factors aggregates several vectors of type factor into a single one to allow multidimensional ipu-steps. See examples.

Usage

ipf_step_ref(w, classes, targets)

ipf_step(w, classes, targets)

ipf_step_f(w, classes, targets)

combine_factors(dat, targets)

Arguments

w

a numeric vector of weights. All entries should be positive.

classes

a factor variable. Must have the same length as w.

targets

key figure to target with the ipu scheme. A numeric verctor of the same length as levels(classes). This can also be a table produced by xtabs. See examples.

dat

a data.frame containing the factor variables to be combined.

Details

ipf_step returns the adjusted weights. ipf_step_ref does the same, but updates w by reference rather than returning. ipf_step_f returns a multiplicator: adjusted weights divided by unadjusted weights. combine_factors is designed to make ipf_step work with contingency tables produced by xtabs.

Examples

############# one-dimensional ipu ##############

## create random data
nobs <- 10
classLabels <- letters[1:3]
dat = data.frame(
  weight = exp(rnorm(nobs)),
  household = factor(sample(classLabels, nobs, replace = TRUE))
)
dat
#>       weight household
#> 1  0.9008129         b
#> 2  0.4867420         c
#> 3  1.9985718         c
#> 4  1.9455129         a
#> 5  0.6525113         c
#> 6  0.4516293         a
#> 7  2.2275627         a
#> 8  1.4814361         b
#> 9  1.0449447         c
#> 10 1.2768816         a

## create targets (same lenght as classLabels!)
targets <- 3:5

## calculate weights
new_weight <- ipf_step(dat$weight, dat$household, targets)
cbind(dat, new_weight)
#>       weight household new_weight
#> 1  0.9008129         b  1.5125420
#> 2  0.4867420         c  0.5818417
#> 3  1.9985718         c  2.3890530
#> 4  1.9455129         a  0.9889779
#> 5  0.6525113         c  0.7799991
#> 6  0.4516293         a  0.2295803
#> 7  2.2275627         a  1.1323545
#> 8  1.4814361         b  2.4874580
#> 9  1.0449447         c  1.2491062
#> 10 1.2768816         a  0.6490873

## check solution
xtabs(new_weight ~ dat$household)
#> dat$household
#> a b c 
#> 3 4 5 

## calculate weights "by reference"
ipf_step_ref(dat$weight, dat$household, targets)
dat
#>       weight household
#> 1  1.5125420         b
#> 2  0.5818417         c
#> 3  2.3890530         c
#> 4  0.9889779         a
#> 5  0.7799991         c
#> 6  0.2295803         a
#> 7  1.1323545         a
#> 8  2.4874580         b
#> 9  1.2491062         c
#> 10 0.6490873         a

############# multidimensional ipu ##############

## load data
factors <- c("time", "sex", "smoker", "day")
tips <- data.frame(sex=c("Female","Male","Male"), day=c("Sun","Mon","Tue"),
time=c("Dinner","Lunch","Lunch"), smoker=c("No","Yes","No"))
tips <- tips[factors]

## combine factors
con <- xtabs(~., tips)
cf <- combine_factors(tips, con)
cbind(tips, cf)[sample(nrow(tips), 10, replace = TRUE),]
#>      time  sex smoker day cf
#> 3   Lunch Male     No Tue 20
#> 2   Lunch Male    Yes Mon  8
#> 3.1 Lunch Male     No Tue 20
#> 3.2 Lunch Male     No Tue 20
#> 2.1 Lunch Male    Yes Mon  8
#> 3.3 Lunch Male     No Tue 20
#> 3.4 Lunch Male     No Tue 20
#> 2.2 Lunch Male    Yes Mon  8
#> 2.3 Lunch Male    Yes Mon  8
#> 2.4 Lunch Male    Yes Mon  8

## adjust weights
weight <- rnorm(nrow(tips)) + 5
adjusted_weight <- ipf_step(weight, cf, con)

## check outputs
con2 <- xtabs(adjusted_weight ~ ., data = tips)
sum((con - con2)^2)
#> [1] 0