Perform one step of iterative proportional updating

C++ routines to invoke a single iteration of the Iterative proportional updating (IPU) scheme. Targets and classes are assumed to be one dimensional in the ipf_step functions. combine_factors aggregates several vectors of type factor into a single one to allow multidimensional ipu-steps. See examples.

ipf_step_ref(w, classes, targets)

ipf_step(w, classes, targets)

ipf_step_f(w, classes, targets)

combine_factors(dat, targets)

Arguments

w
a numeric vector of weights. All entries should be positive.

classes
a factor variable. Must have the same length as w.

targets
key figure to target with the ipu scheme. A numeric verctor of the same length as levels(classes). This can also be a table produced by xtabs. See examples.

dat
a data.frame containing the factor variables to be combined.

Details

ipf_step returns the adjusted weights. ipf_step_ref does the same, but updates w by reference rather than returning. ipf_step_f returns a multiplicator: adjusted weights divided by unadjusted weights. combine_factors is designed to make ipf_step work with contingency tables produced by xtabs.

Examples

############# one-dimensional ipu ##############

## create random data
nobs <- 10
classLabels <- letters[1:3]
dat = data.frame(
  weight = exp(rnorm(nobs)),
  household = factor(sample(classLabels, nobs, replace = TRUE))
)
dat

#>       weight household
#> 1  0.5205577         c
#> 2  2.6770281         a
#> 3  7.0644112         b
#> 4  3.4453868         b
#> 5  3.8267252         a
#> 6  0.8028524         b
#> 7  0.7911008         a
#> 8  6.2139009         a
#> 9  0.6293973         c
#> 10 1.2891310         b

## create targets (same lenght as classLabels!)
targets <- 3:5

## calculate weights
new_weight <- ipf_step(dat$weight, dat$household, targets)
cbind(dat, new_weight)

#>       weight household new_weight
#> 1  0.5205577         c  2.2633829
#> 2  2.6770281         a  0.5945096
#> 3  7.0644112         b  2.2423532
#> 4  3.4453868         b  1.0936190
#> 5  3.8267252         a  0.8498322
#> 6  0.8028524         b  0.2548378
#> 7  0.7911008         a  0.1756862
#> 8  6.2139009         a  1.3799719
#> 9  0.6293973         c  2.7366171
#> 10 1.2891310         b  0.4091901

## check solution
xtabs(new_weight ~ dat$household)

#> dat$household
#> a b c 
#> 3 4 5 

## calculate weights "by reference"
ipf_step_ref(dat$weight, dat$household, targets)
dat

#>       weight household
#> 1  2.2633829         c
#> 2  0.5945096         a
#> 3  2.2423532         b
#> 4  1.0936190         b
#> 5  0.8498322         a
#> 6  0.2548378         b
#> 7  0.1756862         a
#> 8  1.3799719         a
#> 9  2.7366171         c
#> 10 0.4091901         b

############# multidimensional ipu ##############

## load data
factors <- c("time", "sex", "smoker", "day")
tips <- data.frame(sex=c("Female","Male","Male"), day=c("Sun","Mon","Tue"),
time=c("Dinner","Lunch","Lunch"), smoker=c("No","Yes","No"))
tips <- tips[factors]

## combine factors
con <- xtabs(~., tips)
cf <- combine_factors(tips, con)
cbind(tips, cf)[sample(nrow(tips), 10, replace = TRUE),]

#>       time    sex smoker day cf
#> 3    Lunch   Male     No Tue 20
#> 2    Lunch   Male    Yes Mon  8
#> 2.1  Lunch   Male    Yes Mon  8
#> 1   Dinner Female     No Sun  9
#> 3.1  Lunch   Male     No Tue 20
#> 1.1 Dinner Female     No Sun  9
#> 3.2  Lunch   Male     No Tue 20
#> 2.2  Lunch   Male    Yes Mon  8
#> 3.3  Lunch   Male     No Tue 20
#> 3.4  Lunch   Male     No Tue 20

## adjust weights
weight <- rnorm(nrow(tips)) + 5
adjusted_weight <- ipf_step(weight, cf, con)

## check outputs
con2 <- xtabs(adjusted_weight ~ ., data = tips)
sum((con - con2)^2)

#> [1] 0